rename qconfig to config; remove some bot warnings
